optimalizating CRUD operations

+++ App/Controllers/PizzaController.php
@@ -5,6 +5,7 @@ namespace App\Controllers;
 use App\Core\AControllerBase;
 use App\Core\Responses\Response;
 use App\Models\Pizza;
+use http\Header\Parser;
 
 class PizzaController extends AControllerBase
 {
@@ -30,14 +31,58 @@ class PizzaController extends AControllerBase
         $cost = $formData->getValue("cost");
         $imagePath = "/public/images/pizzas/neapolitan.png";
 
-        $pizza = new Pizza();
-        $pizza->setName($name);
-        $pizza->setDescription($description);
-        $pizza->setCost($cost);
-        $pizza->setImagePath($imagePath);
-        $pizza->save();
+        if ($this->validateInputOnAdd($name, $description, $cost, $imagePath)) {
+            $data = [];
+            $data["message"] = "Item with successfully added!";
 
-        return $this->redirect($this->url('shop.index'));
+            $pizza = new Pizza();
+            $pizza->setName($name);
+            $pizza->setDescription($description);
+            $pizza->setCost($cost);
+            $pizza->setImagePath($imagePath);
+            $pizza->save();
+        } else {
+            $data["message"] = "Invalid input values!";
+        }
+
+        return $this->redirect($this->url("shop.add", ["data" => $data]));
+    }
+
+    public function update(): Response
+    {
+        return $this->html();
+    }
+
+    /**
+     * @throws \Exception
+     */
+    public function updateItem(): Response
+    {
+        $formData = $this->app->getRequest();
+        $id = $formData->getValue("id");
+        $name = $formData->getValue("name");
+        $description = $formData->getValue("description");
+        $cost = $formData->getValue("cost");
+        $imagePath = "/public/images/pizzas/neapolitan.png";
+
+        if ($this->validateInputOnUpdate($id, $name, $description, $cost, $imagePath)) {
+            $data = [];
+            $pizzaGetOne = Pizza::getOne($id);
+
+            if (is_null($pizzaGetOne)) {
+                $data["message"] = "Item with id (" . $id . ") doesnt exists!";
+            } else {
+                $data["message"] = "Item with id (" . $id . ") successfully updated!";
+                $pizzaGetOne->setName($name);
+                $pizzaGetOne->setDescription($description);
+                $pizzaGetOne->setCost($cost);
+                $pizzaGetOne->save();
+            }
+        } else {
+            $data["message"] = "Invalid input values!";
+        }
+
+        return $this->redirect($this->url("shop.update", ["data" => $data]));
     }
 
     public function remove(): Response
@@ -48,10 +93,44 @@ class PizzaController extends AControllerBase
     public function removeItem(): Response
     {
         $formData = $this->app->getRequest();
-        $id = $formData->getValue('id');
+        $id = $formData->getValue("id");
         $pizzaGetOne = Pizza::getOne($id);
-        $pizzaGetOne->delete();
 
-        return $this->redirect($this->url('shop.remove'));
+        if ($this->validateInputOnRemove($id)) {
+            $data = [];
+            if (is_null($pizzaGetOne)) {
+                $data["message"] = "Item with id (" . $id . ") doesnt exist!";
+            } else {
+                $data["message"] = "Item with id (" . $id . ") successfully removed!";
+                $pizzaGetOne->delete();
+                $pizzaGetOne->save();
+            }
+        } else {
+            $data["message"] = "Invalid input values!";
+        }
+
+        return $this->redirect($this->url("shop.remove", ["data" => $data]));
+    }
+
+    public function validateInputOnAdd($name, $description, $cost, $imagePath): bool
+    {
+        return strlen($name) > 0 && strlen($name) < 50 &&
+            strlen($description) > 0 && strlen($description) < 300 &&
+            !is_null($cost) && strlen($cost) > 0 && is_int((int)$cost) &&
+            strlen($imagePath) > 0 && strlen($imagePath) < 300;
+    }
+
+    public function validateInputOnUpdate($id, $name, $description, $cost, $imagePath): bool
+    {
+        return !is_null($id) && strlen($id) > 0 && is_int((int)$id) &&
+            strlen($name) > 0 && strlen($name) < 50 &&
+            strlen($description) > 0 && strlen($description) < 300 &&
+            !is_null($cost) && strlen($cost) > 0 && is_int((int)$cost) &&
+            strlen($imagePath) > 0 && strlen($imagePath) < 300;
+    }
+
+    public function validateInputOnRemove($id): bool
+    {
+        return is_int((int)$id);
     }
 }
